backend/pyproject.toml
[project]
name = "realtime-assistant-backend"
version = "0.1.0"
description = "Real-time screen + voice + text assistant backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "aiortc>=1.9.0",
    "av>=12.3.0",
    "numpy>=1.26.0",
    "Pillow>=10.4.0",
    "imagehash>=4.3.1",
    "websockets>=13.0",
    "httpx>=0.27.0",
    "azure-cognitiveservices-speech>=1.40.0",
]

[project.optional-dependencies]
m4 = [
    "openai>=1.54.0",
    "agent-framework>=0.1.0",
]
m5 = [
    "playwright>=1.47.0",
]
m6 = [
    "azure-cosmos>=4.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
